<Fix> Routing 책임 분리 (app -> router.tsx)
