+++ Todo/최원빈/TODOCHOI/src/main.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+
 import './index.css';
-import { BrowserRouter } from 'react-router-dom';
+import { router } from './router/router';
+import { RouterProvider } from 'react-router-dom';
+
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
+    <RouterProvider router={router} />
   </React.StrictMode>
 );
